+++ src/core/async-work-cache.ts
@@ -12,9 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {CancelToken, isCancel} from 'cancel-token';
-
-const neverCancels = CancelToken.source().token;
+import {isCancel} from 'cancel-token';
+import {MinimalCancelToken, neverCancels} from './cancel-token';
 
 /**
  * A map from keys to promises of values. Used for caching asynchronous work.
@@ -39,13 +38,13 @@ export class AsyncWorkCache<K, V> {
    * to compute the value for `key` once, no matter how often or with what
    * timing getOrCompute is called, even recursively.
    *
-   * If a compute() throws a Cancel that should not result in a Cancel for
-   * non-cancelled operations. So long as the given cancelToken is not
-   * cancelled, the caller will not receive a Cancel.
+   * If one call to getOrCompute is cancelled that that should not result in a
+   * Cancel for non-cancelled operations. So long as the given cancelToken is
+   * not cancelled, the caller will not receive a Cancel.
    */
   async getOrCompute(
-      key: K, compute: () => Promise<V>, cancelToken?: CancelToken) {
-    cancelToken = cancelToken || neverCancels;
+      key: K, compute: () => Promise<V>,
+      cancelToken: MinimalCancelToken = neverCancels) {
     cancelToken.throwIfRequested();
     while (true) {
       try {
@@ -55,6 +54,8 @@ export class AsyncWorkCache<K, V> {
       } catch (err) {
         cancelToken.throwIfRequested();
         if (isCancel(err)) {
+          // Ok, whoever was working on computing `key` was cancelled, but it
+          // wasn't us because we're not cancelled. Let's try again!.
           continue;
         }
         throw err;